requirements.txt
# Web框架
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# 数据处理
pandas>=2.1.3
numpy>=1.26.0
polars>=0.19.19
pyarrow>=14.0.1

# 数据库
sqlalchemy>=2.0.23
alembic>=1.12.1
psycopg2-binary>=2.9.9
redis>=5.0.1
influxdb-client>=1.38.0

# 机器学习
scikit-learn>=1.3.2

# 技术分析与优化
pandas-ta>=0.3.14b0
optuna>=3.4.0

# 异步任务
celery>=5.3.4
flower>=2.0.1

# 数据采集
ccxt>=4.1.38
yfinance>=0.2.24
requests>=2.31.0
aiohttp>=3.9.1

# 交易执行
ib_insync>=0.9.86

# 配置管理
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.1

# 日志和监控
loguru>=0.7.2
prometheus-client>=0.19.0

# 测试
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
httpx>=0.25.2

# 可视化与报告
plotly>=5.17.0
matplotlib>=3.8.2
seaborn>=0.13.0
jinja2>=3.1.2

# 统计分析
scipy>=1.11.4

# 其他工具
click>=8.1.7
rich>=13.7.0
schedule>=1.2.0
python-dateutil>=2.8.2
pytz>=2023.3
